fixed #9 add pester tests that run the script analyzer rules (#37)
* fixed #9 add pester tests that run the script analyzer rules

* Update ScriptAnalyzer.tests.ps1

* fixed script analyzer rule warnings

* remove script analyzer for ps1 files, remove rule PSUseSingularNoun, Fix naming of Copy-PackageContent

* added support for whatif in New-TestSolution

* added output type for script analyzsis

+++ TestContent/TestSolution/New-TestSolution.psm1
@@ -1,15 +1,17 @@
 Import-Module "$PSScriptRoot\..\..\Get-MSBuild\Get-MSBuild.psm1" -Force
 
 Function New-TestSolution {
+    [CmdletBinding(SupportsShouldProcess = $true)]
+    [OutputType([System.String])]            
     Param(
         [Parameter(Mandatory = $True)]
         [string]$TempPath
     )
 
     $tempTestSolutionPath = "$TempPath\TestSolution"
-    
-    Remove-Item -Path "$tempTestSolutionPath" -Recurse -Force -ErrorAction SilentlyContinue
-
+    if ($pscmdlet.ShouldProcess($tempTestSolutionPath, "Removing all files in folder")) {
+        Remove-Item -Path "$tempTestSolutionPath" -Recurse -Force -ErrorAction SilentlyContinue
+    }
     # Copy solution
     Copy-Item "$PSScriptRoot\src" -Destination "$tempTestSolutionPath\src" -Container -Recurse
     Copy-Item "$PSScriptRoot\TestSolution.sln" -Destination "$tempTestSolutionPath"
